initial work to clean up ffi module
