+++ src/ffi/cpython/ceval.rs
@@ -1,0 +1,19 @@
+#[cfg(not(Py_LIMITED_API))]
+use crate::ffi::code::FreeFunc;
+use crate::ffi::object::PyObject;
+use crate::ffi::pystate::Py_tracefunc;
+use std::os::raw::c_int;
+
+pub type _PyFrameEvalFunction =
+    extern "C" fn(*mut crate::ffi::PyFrameObject, c_int) -> *mut PyObject;
+
+extern "C" {
+    pub fn _PyEval_EvalFrameDefault(
+        arg1: *mut crate::ffi::PyFrameObject,
+        exc: c_int,
+    ) -> *mut PyObject;
+    #[cfg(not(Py_LIMITED_API))]
+    pub fn _PyEval_RequestCodeExtraIndex(func: FreeFunc) -> c_int;
+    pub fn PyEval_SetProfile(trace_func: Py_tracefunc, arg1: *mut PyObject);
+    pub fn PyEval_SetTrace(trace_func: Py_tracefunc, arg1: *mut PyObject);
+}
