+++ src/ffi/cpython/mod.rs
@@ -1,3 +1,7 @@
 pub mod abstract_;
+pub mod bytesobject;
+pub mod ceval;
 
 pub use self::abstract_::*;
+pub use self::bytesobject::*;
+pub use self::ceval::*;
